1.18.1 release NRLMMD-GEOIPS/geoips#1246
