--- docs/source/releases/latest/update-black-formatting.yaml
@@ -1,15 +1,0 @@
-continuous integration:
-- title: 'Update code base for Black 2026 style - see Black 26.1.0 release notes for details.'
-  description: 'Please make sure you update your local Black install as old versions will produce different results.'
-  files:
-    added: []
-    modified:
-    - 'my_package/plugins/modules/algorithms/pmw_89test.py'
-    deleted: []
-    moved: []
-  related-issue:
-    number: null
-    repo_url: ''
-  date:
-    start: Jan 20, 2026
-    finish: Jan 20, 2026
